+++ modules/global-catalog/src/test/java/com/ibm/cloud/platform_services/global_catalog/v1/model/GetChildObjectsOptionsTest.java
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright IBM Corp. 2020.
+ * (C) Copyright IBM Corp. 2021.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -40,7 +40,9 @@ public class GetChildObjectsOptionsTest {
       .sortBy("testString")
       .descending("testString")
       .languages("testString")
-      .complete("testString")
+      .complete(true)
+      .offset(Long.valueOf("26"))
+      .limit(Long.valueOf("200"))
       .build();
     assertEquals(getChildObjectsOptionsModel.id(), "testString");
     assertEquals(getChildObjectsOptionsModel.kind(), "testString");
@@ -50,7 +52,9 @@ public class GetChildObjectsOptionsTest {
     assertEquals(getChildObjectsOptionsModel.sortBy(), "testString");
     assertEquals(getChildObjectsOptionsModel.descending(), "testString");
     assertEquals(getChildObjectsOptionsModel.languages(), "testString");
-    assertEquals(getChildObjectsOptionsModel.complete(), "testString");
+    assertEquals(getChildObjectsOptionsModel.complete(), Boolean.valueOf(true));
+    assertEquals(getChildObjectsOptionsModel.offset(), Long.valueOf("26"));
+    assertEquals(getChildObjectsOptionsModel.limit(), Long.valueOf("200"));
   }
 
   @Test(expectedExceptions = IllegalArgumentException.class)